feat: show differnt success message for password on success, refactor
