+++ cypress/e2e/SignIn.cy.ts
@@ -10,7 +10,7 @@ describe('Name and Email Validation', () => {
     cy.visit(SIGN_IN_PATH);
     // Signing in with a wrong email format
     cy.signInAndCheck(WRONG_EMAIL);
-    // Siging in with a valid email
+    // Signing in with a valid email
     cy.signInAndCheck(GRAASP);
   });
 });
