+++ cypress/e2e/SignInPassword.cy.ts
@@ -1,6 +1,7 @@
 import { API_ROUTES } from '@graasp/query-client';
 
 import { SIGN_IN_PATH } from '../../src/config/paths';
+import { PASSWORD_SUCCESS_ALERT } from '../../src/config/selectors';
 import { MEMBERS } from '../fixtures/members';
 
 describe('Email and Password Validation', () => {
@@ -27,5 +28,29 @@ describe('Email and Password Validation', () => {
     cy.signInPasswordAndCheck(GRAASP);
 
     cy.url().should('contain', redirectionLink);
+  });
+
+  it('Sign In With Password shows success message if no redirect', () => {
+    cy.intercept(
+      {
+        pathname: API_ROUTES.SIGN_IN_WITH_PASSWORD_ROUTE,
+      },
+      (req) => {
+        req.reply({ statusCode: 303 });
+      },
+    ).as('signInWithPassword');
+
+    const { WRONG_EMAIL, WRONG_PASSWORD, GRAASP } = MEMBERS;
+    cy.visit(SIGN_IN_PATH);
+    // Select sign in method
+    cy.signInPasswordMethodAndCheck();
+    // Signing in with wrong email
+    cy.signInPasswordAndCheck(WRONG_EMAIL);
+    // Signing in with a valid email but empty password
+    cy.signInPasswordAndCheck(WRONG_PASSWORD);
+    // Signing in with a valid email and password
+    cy.signInPasswordAndCheck(GRAASP);
+
+    cy.get(`#${PASSWORD_SUCCESS_ALERT}`).should('be.visible');
   });
 });
