+++ src/App.tsx
@@ -2,6 +2,7 @@ import * as Sentry from '@sentry/react';
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
 
 import ErrorFallback from './components/ErrorFallback';
+import MagicLinkSuccessContent from './components/MagicLinkSuccessContent';
 import MobileAuth from './components/MobileAuth';
 import Redirection from './components/Redirection';
 import SignIn from './components/SignIn';
@@ -9,6 +10,7 @@ import SignUp from './components/SignUp';
 import {
   HOME_PATH,
   MOBILE_AUTH_PATH,
+  SIGN_IN_MAGIC_LINK_SUCCESS_PATH,
   SIGN_IN_PATH,
   SIGN_UP_PATH,
 } from './config/paths';
@@ -19,6 +21,10 @@ const App = () => (
       <Redirection>
         <Routes>
           <Route path={SIGN_IN_PATH} element={<SignIn />} />
+          <Route
+            path={SIGN_IN_MAGIC_LINK_SUCCESS_PATH}
+            element={<MagicLinkSuccessContent />}
+          />
           <Route path={SIGN_UP_PATH} element={<SignUp />} />
           <Route path={HOME_PATH} element={<SignIn />} />
           <Route path={MOBILE_AUTH_PATH} element={<MobileAuth />} />
