+++ src/langs/constants.ts
@@ -40,4 +40,5 @@ export const AUTH = {
   INVALID_EMAIL_ERROR: 'INVALID_EMAIL_ERROR',
   EMPTY_EMAIL_ERROR: 'EMPTY_EMAIL_ERROR',
   PASSWORD_EMPTY_ERROR: 'PASSWORD_EMPTY_ERROR',
+  PASSWORD_SUCCESS_ALERT: 'PASSWORD_SUCCESS_ALERT',
 };
