+++ src/langs/en.json
@@ -38,5 +38,6 @@
   "USERNAME_TOO_LONG_ERROR": "Please enter a username under {{max}} characters",
   "INVALID_EMAIL_ERROR": "This does not look like a valid email address",
   "EMPTY_EMAIL_ERROR": "An email address is required, this field can not be empty",
-  "PASSWORD_EMPTY_ERROR": "The password can not be empty"
+  "PASSWORD_EMPTY_ERROR": "The password can not be empty",
+  "PASSWORD_SUCCESS_ALERT": "You successfully signed in. You will be redirected soon."
 }
